Resolve MIGRATE-ACTION, improve dependencies, plugins and modern CMake use
* Improve header includes / forward declarations to reduce transitive dependencies.

* Split plugin registration source from implementation.

* New INTERFACE targets:

  * webevd::ThreadsafeGalleryEvent

  * webevd::JSONFormatter

* Use most appropriate plugin builders.

* Exported variable `webevd_WEB_DIR` pointing to installed location of
  web support files (CSS, etc.)
