+++ webevd/CMakeLists.txt
@@ -1,4 +1,2 @@
-set( IFDH_ART_DIR $ENV{IFDH_ART_DIR} )
-
-add_subdirectory(WebEVD)
 add_subdirectory(ArtSupport)
+add_subdirectory(WebEVD)
