+++ webevd/WebEVD/gallery/CMakeLists.txt
@@ -1,19 +1,14 @@
-set( evd_lib_list pthread
-                  canvas::canvas
-                  cetlib::cetlib
-                  cetlib_except::cetlib_except
-                  gallery::gallery
-                  nusimdata::SimulationBase
-                  lardataobj_RecoBase
-                  webevd_ArtSupport
-                  png
-                  webevd_WebEVD
-                  )
-
-cet_make_exec( webevd
-               SOURCE webevd.cxx
-               LIBRARIES ${evd_lib_list}
-               )
+cet_make_exec(webevd SOURCE webevd.cxx
+  LIBRARIES PRIVATE
+  webevd::WebEVD
+  webevd::ThreadsafeGalleryEvent
+  webevd::ArtSupport
+  lardata::DetectorPropertiesService
+  larcore::Geometry_Geometry_service
+  gallery::gallery
+  art::Framework_Services_Registry
+  canvas::canvas
+)
 
 install_headers()
 install_source()
